requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.1
cryptography==41.0.7
email-validator==2.1.0
secrets
hashlib
time
json
re
os
base64